// FE-navigation3/src/app/api/fn/admin/admin-upload-sample-image-data-post-plain.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
import { Observable } from 'rxjs';
import { filter, map } from 'rxjs/operators';
import { StrictHttpResponse } from '../../strict-http-response';
import { RequestBuilder } from '../../request-builder';


export interface AdminUploadSampleImageDataPost$Plain$Params {
      body?: {
'AnnotationFile'?: Blob;
'SampleImageFileName'?: string;
'OrganId'?: number;
'KeyWords'?: string;
}
}

export function adminUploadSampleImageDataPost$Plain(http: HttpClient, rootUrl: string, params?: AdminUploadSampleImageDataPost$Plain$Params, context?: HttpContext): Observable<StrictHttpResponse<string>> {
  const rb = new RequestBuilder(rootUrl, adminUploadSampleImageDataPost$Plain.PATH, 'post');
  if (params) {
    rb.body(params.body, 'multipart/form-data');
  }

  return http.request(
    rb.build({ responseType: 'text', accept: 'text/plain', context })
  ).pipe(
    filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
    map((r: HttpResponse<any>) => {
      return r as StrictHttpResponse<string>;
    })
  );
}

adminUploadSampleImageDataPost$Plain.PATH = '/admin/upload-sample-image-data';